refactor(net,tunnel): group the feature's new code into module hierarchies
The port-mapping protocol code was scattered across ~7 top-level net/
files and the tunnel's forwarding backend across 3. Group them:

- net/port_map/  — client.rs (controller), upnp.rs, pcp/{hostname,portset},
  server/{mod.rs (GatewayBackend+handle), igd.rs, igd_xml/}
- net/dns_update/ — mod.rs + rfc2136.rs
- tunnel/forward/ — sni.rs, pcp.rs, igd.rs

Pure reorg: file moves + import-path updates, no behavior change. lib
compiles, 107 tests pass.
