+++ core/src/net/mod.rs
@@ -7,22 +7,16 @@ pub mod forward;
 pub mod gateway;
 pub mod host;
 pub mod http;
-pub mod igd_server;
 pub mod keys;
 pub mod mdns;
 pub mod net_controller;
-pub mod pcp_hostname;
-pub mod pcp_portset;
-pub mod pcp_server;
 pub mod port_map;
-pub mod rfc2136;
 pub mod service_interface;
 pub mod socks;
 pub mod ssl;
 pub mod static_server;
 pub mod tls;
 pub mod tunnel;
-pub mod upnp;
 pub mod utils;
 pub mod vhost;
 pub mod web_server;
