+++ core/src/net/port_map/mod.rs
@@ -1,0 +1,10 @@
+//! NAT port-mapping protocols: a client that asks an upstream gateway to open
+//! ports (PCP/NAT-PMP/UPnP) and a reusable server that answers such requests
+//! (see [`server`]).
+
+mod client;
+pub mod pcp;
+pub mod server;
+pub mod upnp;
+
+pub use client::{PortMapController, candidate_gateways};
