+++ core/src/net/port_map/pcp/mod.rs
@@ -1,0 +1,5 @@
+//! PCP option extensions shared by client and server: HOSTNAME (SNI demux) and
+//! PORT_SET (contiguous port ranges).
+
+pub mod hostname;
+pub mod portset;
