+++ core/src/net/port_map/server/igd.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 use axum::http::{HeaderMap, StatusCode, header};
 use axum::response::{IntoResponse, Response};
 
-use crate::net::pcp_server::GatewayBackend;
+use crate::net::port_map::server::GatewayBackend;
 
 pub const SSDP_MULTICAST: Ipv4Addr = Ipv4Addr::new(239, 255, 255, 250);
 pub const SSDP_PORT: u16 = 1900;
