+++ core/src/tunnel/mod.rs
@@ -10,10 +10,8 @@ pub mod auth;
 pub mod context;
 pub mod db;
 pub mod dns;
-pub mod igd;
+pub mod forward;
 pub(crate) mod migrations;
-pub mod pcp;
-pub mod sni;
 pub mod update;
 pub mod web;
 pub mod wg;
